+++ lib/searchkick/index.rb
@@ -477,7 +477,7 @@ module Searchkick
         end
 
         mapping_options = Hash[
-          [:autocomplete, :suggest, :word, :text_start, :text_middle, :text_end, :word_start, :word_middle, :word_end, :highlight, :searchable, :only_analyzed]
+          [:autocomplete, :suggest, :word, :text_start, :text_middle, :text_end, :word_start, :word_middle, :word_end, :highlight, :searchable, :filterable, :only_analyzed]
             .map { |type| [type, (options[type] || []).map(&:to_s)] }
         ]
 
@@ -486,7 +486,7 @@ module Searchkick
         mapping_options.values.flatten.uniq.each do |field|
           fields = {}
 
-          if mapping_options[:only_analyzed].include?(field)
+          if mapping_options[:only_analyzed].include?(field) || (options.key?(:filterable) && !mapping_options[:filterable].include?(field))
             fields[field] = {type: default_type, index: "no"}
           else
             fields[field] = keyword_mapping
@@ -501,7 +501,7 @@ module Searchkick
               end
             end
 
-            mapping_options.except(:highlight, :searchable, :only_analyzed, :word).each do |type, f|
+            mapping_options.except(:highlight, :searchable, :filterable, :only_analyzed, :word).each do |type, f|
               if options[:match] == type || f.include?(field)
                 fields[type] = {type: default_type, index: "analyzed", analyzer: "searchkick_#{type}_index"}
               end
@@ -547,6 +547,10 @@ module Searchkick
           # and the _all index analyzer will take care of it
           "{name}" => keyword_mapping.merge(include_in_all: !options[:searchable])
         }
+
+        if options.key?(:filterable)
+          dynamic_fields["{name}"] = {type: default_type, index: "no"}
+        end
 
         dynamic_fields["{name}"][:ignore_above] = 256 unless below22