+++ test/test_helper.rb
@@ -287,9 +287,10 @@ class Product
     word_middle: [:name],
     word_end: [:name],
     highlight: [:name],
-    # unsearchable: [:description],
     searchable: [:name, :color],
-    only_analyzed: [:alt_description],
+    filterable: [:name, :color, :description],
+    # unsearchable: [:description],
+    # only_analyzed: [:alt_description],
     match: ENV["MATCH"] ? ENV["MATCH"].to_sym : nil
 
   attr_accessor :conversions, :user_ids, :aisle, :details